feature/audio-preview: return back models int store state file
