+++ src/app/modules/editor/components/editor-layout/editor-layout.component.ts
@@ -20,7 +20,8 @@ export class EditorLayoutComponent implements OnInit {
 
   private editorState: Observable<EditorState> =
     this.store.select(selectEditor);
-  gameFolder: string = 'C:\\Users\\ctbg-computer\\Desktop\\мемные игры\\тест';
+
+  gameFolder: string = '/Users/vlad3k/Desktop/papka';
   public gameName: any;
   public mode: 'file' | 'slide' = 'slide';
