+++ src/app/modules/game-play/store/game-play.state.ts
@@ -1,6 +1,5 @@
-import { EditorState } from '../../editor/state/editor.state';
+import { EditorState, Slide } from '../../editor/state/editor.state';
 import { PlayerEntity } from '../players-store/players.state';
-import { Slide } from '../../../models/models';
 
 export interface GamePlayState {
   allGame: EditorState;
