+++ src/app/shared/utils/type-guards.ts
@@ -1,4 +1,4 @@
-import { QuestionWithAudioSlide, QuestionWithImageSlide, Slide } from '../../models/models';
+import { QuestionWithAudioSlide, QuestionWithImageSlide, Slide } from '../../modules/editor/state/editor.state';
 
 export const isQuestionWithImageSlide = (slide: Slide): slide is QuestionWithImageSlide => slide.type === 'questionWithImage';
 export const isQuestionWithAudioSlide = (slide: Slide): slide is QuestionWithAudioSlide => slide.type === 'questionWithAudio';
